+++ src/components/homePage/assignments/staff/Resources.js
@@ -1,6 +1,6 @@
 import React, { useState, useContext } from 'react';
 import { Table, Spin } from "antd";
-import ModalAdd from '../../dashboard/staffDashboard/modalAdd/ModalAdd';
+import MultiPurposeModal from '../../dashboard/staffDashboard/multiPurposeModal/MultiPurposeModal';
 import { Input } from "antd";
 import UserContext from '../../../contexts/UserContext';
 
@@ -63,7 +63,7 @@ const Resources = ({ resources }) => {
             {
                 resources.length ? (
                     <>
-                        <ModalAdd 
+                        <MultiPurposeModal 
                             handleOk={ handleOk }
                             addTitle="Add Resource"
                         >
@@ -93,7 +93,7 @@ const Resources = ({ resources }) => {
                                 {/* <Button type="primary" htmlType="submit" >
                                     Create Course
                                 </Button> */}
-                        </ModalAdd>
+                        </MultiPurposeModal>
                         <Table
                             columns={columns}
                             dataSource={data}
