+++ src/components/homePage/staff/createClass/CreateClass.js
@@ -11,7 +11,7 @@ import {
 import * as ROUTES from "../../../../constants/routes";
 import { StyledSectionStaff, StyledDiv } from "../../assignments/styles";
 import CreateAssignments from '../../assignments/staff/CreateAssignments';
-import Courses from '../../assignments/staff/Courses';
+import Courses from '../../assignments/staff/courses/Courses';
 import Units from '../../assignments/staff/Units';
 import Lessons from '../../assignments/staff/Lessons';
 import Resources from '../../assignments/staff/Resources';
